refactoring for new backend
