+++ med-cabinet8/src/components/Settings.js
@@ -1,15 +1,79 @@
-import React from 'react';
+import React, { useState, useEffect } from "react";
+import { useHistory } from 'react-router-dom';
+import axiosWithAuth from "../utils/axiosWithAuth";
 
-const Settings = () => {
-	return (
-		<div>
-			<h1>Settings</h1>
-			<ul>
-				<li>Delete</li>
-				<li>Update </li>
-			</ul>
-		</div>
-	);
+
+
+export const Profile = () =>  {
+const { push } = useHistory();
+const [message, setMessage] = useState('');
+const [password, setPassword] = useState('');
+
+const getData = () => {
+  axiosWithAuth()
+    .get("/profile")
+    .then((res) => {
+      //  console.log("ea: Recommendations.js getData results:", res.data);
+      setMessage(
+        res.data.message,
+      );
+    })
+    .catch((err) =>
+      console.error(
+        "ea: Profile.js: getData: err.message: ",
+        err.message
+      )
+    );
 };
 
-export default Settings;
+useEffect(()=>{
+  getData()
+},[])
+
+  
+
+  const deleteUser = () => {
+    axiosWithAuth()
+      .delete("/profile/delete-user")
+      .then((res) => {
+        console.log(res.data)
+        localStorage.removeItem('token');
+        localStorage.removeItem("login");
+        push('/');
+      })
+      .catch((err) =>
+        console.error(
+          "ea: Profile.js: delete-user: err.message: ",
+          err.results
+        )
+      );
+  };
+
+  const changePassword = () => {
+    const pWord = {"password": password}
+    axiosWithAuth()
+      .put("/profile/change-password", pWord)
+      .then((res) => {
+        localStorage.removeItem('token');
+        localStorage.removeItem("login");
+        push('/');
+      })
+      .catch((err) =>
+        console.error(
+          "ea: Profile.js: chg-password: err.message: ",
+          err.results
+        )
+      );
+  };
+
+
+    return (
+      <div>
+        <p>TODO ADD form to change password</p>
+        <button onClick={deleteUser}>Delete Profile</button>
+      </div>
+    );
+  }
+
+
+export default Profile;
